+++ merchant-persistence/src/main/java/com/mcipay/persistence/mapper/PaymentJournalEntityMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface PaymentJournalEntityMapper {
-    int countByExample(PaymentJournalEntityCriteria example);
+    long countByExample(PaymentJournalEntityCriteria example);
 
     int deleteByExample(PaymentJournalEntityCriteria example);
